split up training utils
